# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Payment Sandbox API",
    "version": "1.0.0",
    "description": "HTTP API for deterministic payment integration testing."
  },
  "servers": [{"url": "http://127.0.0.1:8080"}],
...
  "paths": {
...
    "/payments": {
      "parameters": [{"$ref": "#/components/parameters/CorrelationID"}],
      "post": {
        "tags": ["Payments"], "summary": "Create a payment", "operationId": "createPayment",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreatePaymentRequest"}, "example": {"id": "demo-payment", "amount": 1000, "currency": "EUR"}}}},
        "responses": {"201": {"description": "Payment created.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Payment"}}}}, "400": {"$ref": "#/components/responses/InvalidRequest"}, "409": {"$ref": "#/components/responses/Conflict"}, "415": {"$ref": "#/components/responses/UnsupportedMediaType"}}
      }
    },
    "/payments/{paymentId}": {
      "parameters": [{"$ref": "#/components/parameters/PaymentId"}, {"$ref": "#/components/parameters/CorrelationID"}],
      "get": {"tags": ["Payments"], "summary": "Retrieve a payment", "operationId": "getPayment", "responses": {"200": {"$ref": "#/components/responses/Payment"}, "404": {"$ref": "#/components/responses/NotFound"}}}
    },
    "/payments/{paymentId}/authorize": {
      "parameters": [{"$ref": "#/components/parameters/PaymentId"}, {"$ref": "#/components/parameters/CorrelationID"}],
      "post": {"tags": ["Payments"], "summary": "Authorize a payment", "operationId": "authorizePayment", "responses": {"200": {"$ref": "#/components/responses/Payment"}, "404": {"$ref": "#/components/responses/NotFound"}, "409": {"$ref": "#/components/responses/Conflict"}}}
    },
    "/payments/{paymentId}/capture": {
      "parameters": [{"$ref": "#/components/parameters/PaymentId"}, {"$ref": "#/components/parameters/CorrelationID"}],
      "post": {"tags": ["Payments"], "summary": "Capture an amount", "operationId": "capturePayment", "requestBody": {"$ref": "#/components/requestBodies/Amount"}, "responses": {"200": {"$ref": "#/components/responses/Payment"}, "400": {"$ref": "#/components/responses/InvalidRequest"}, "404": {"$ref": "#/components/responses/NotFound"}, "409": {"$ref": "#/components/responses/Conflict"}, "415": {"$ref": "#/components/responses/UnsupportedMediaType"}}}
    },
    "/payments/{paymentId}/cancel": {
      "parameters": [{"$ref": "#/components/parameters/PaymentId"}, {"$ref": "#/components/parameters/CorrelationID"}],
      "post": {"tags": ["Payments"], "summary": "Cancel a payment", "operationId": "cancelPayment", "responses": {"200": {"$ref": "#/components/responses/Payment"}, "404": {"$ref": "#/components/responses/NotFound"}, "409": {"$ref": "#/components/responses/Conflict"}}}
    },
    "/payments/{paymentId}/refund": {
      "parameters": [{"$ref": "#/components/parameters/PaymentId"}, {"$ref": "#/components/parameters/CorrelationID"}],
      "post": {"tags": ["Payments"], "summary": "Refund an amount", "operationId": "refundPayment", "requestBody": {"$ref": "#/components/requestBodies/Amount"}, "responses": {"200": {"$ref": "#/components/responses/Payment"}, "400": {"$ref": "#/components/responses/InvalidRequest"}, "404": {"$ref": "#/components/responses/NotFound"}, "409": {"$ref": "#/components/responses/Conflict"}, "415": {"$ref": "#/components/responses/UnsupportedMediaType"}}}
    },
...
  },
  "components": {
    "securitySchemes": {"AdminBearerAuth": {"type": "http", "scheme": "bearer", "description": "The PAYMENT_SANDBOX_ADMIN_TOKEN value."}},
    "parameters": {
      "PaymentId": {"name": "paymentId", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1}},
      "CorrelationID": {"name": "X-Correlation-ID", "in": "header", "required": false, "description": "Optional client correlation identifier. Generated by the API when absent.", "schema": {"type": "string", "maxLength": 128}},
...
    },
    "requestBodies": {
      "Amount": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AmountRequest"}, "example": {"amount": 1000, "currency": "EUR"}}}},
...
    },
    "schemas": {
      "ErrorResponse": {"type": "object", "required": ["error"], "properties": {"error": {"type": "object", "required": ["code", "message"], "properties": {"code": {"type": "string"}, "message": {"type": "string"}}}}},
...
      "Payment": {"type": "object", "required": ["id", "amount", "currency", "status", "authorized_amount", "captured_amount", "refunded_amount", "version"], "properties": {"id": {"type": "string"}, "amount": {"type": "integer", "format": "int64"}, "currency": {"type": "string"}, "status": {"type": "string"}, "authorized_amount": {"type": "integer", "format": "int64"}, "captured_amount": {"type": "integer", "format": "int64"}, "refunded_amount": {"type": "integer", "format": "int64"}, "version": {"type": "integer", "format": "int64"}}},
      "CreatePaymentRequest": {"type": "object", "required": ["id", "amount", "currency"], "additionalProperties": false, "properties": {"id": {"type": "string", "minLength": 1}, "amount": {"type": "integer", "format": "int64", "minimum": 1}, "currency": {"type": "string", "minLength": 1}}},
      "AmountRequest": {"type": "object", "required": ["amount", "currency"], "additionalProperties": false, "properties": {"amount": {"type": "integer", "format": "int64", "minimum": 1}, "currency": {"type": "string", "minLength": 1}}},
...
    },
    "responses": {
...
      "Payment": {"description": "Payment response.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Payment"}}}},
...
      "InvalidRequest": {"description": "Invalid request.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
      "UnsupportedMediaType": {"description": "Content-Type must be application/json.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
...
      "NotFound": {"description": "Requested resource was not found.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
      "Conflict": {"description": "Operation rejected by current state or data.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
    }
  }
}